requirements.txt
appnope==0.1.3
asttokens==2.0.8
autopep8==1.4.4
backcall==0.2.0
bcrypt==4.0.1
blinker==1.4
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
decorator==5.1.1
dnspython==2.2.1
email-validator==1.3.0
executing==1.1.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-DebugToolbar==0.13.1
Flask-SQLAlchemy==3.0.1
Flask-WTF==1.0.1
greenlet==1.1.3.post0
idna==3.4
importlib-metadata==5.0.0
ipython==8.5.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
psycopg2-binary==2.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.5.0
Pygments==2.13.0
python-decouple==3.6
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.41
stack-data==0.5.1
traitlets==5.4.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.9.0
